refactor: TabStorageContext 을 역할에 맞게 파일 위치 변경
